Relocate Handler from app to server
